third_party/blink/web_tests/fast/lists/ol-reversed-nested-list-expected.txt

This tests that list items are properly marked when a reversed list contains nested lists.

10 Ten
9 Nine
8 Eight
7 Seven One Two
-> -> ◦ One
-> -> ◦ Two
6 Six
5 Five
-> 2 Two
-> 1 One
4 Four
3 Three
2 Two
